cdk/requirements.txt
# AWS CDK dependencies
aws-cdk-lib>=2.170.0
constructs>=10.0.0
